requirements.txt
flask
flask_mysqldb
PyMuPDF
openai
python-dotenv
mysql-connector-python
plotly